Update reload API

+++ Credits_Missionlist_Text.user.js
@@ -1,7 +1,7 @@
 // ==UserScript==
 // @name         Credits Missionlist Text
 // @namespace    http://tampermonkey.net/
-// @version      4.0.9
+// @version      4.0.10
 // @description  Credits in Missionlist in Text
 // @author       JRH1997
 // @match        https://www.meldkamerspel.com/
@@ -76,6 +76,13 @@
                 for(var ic2 = 0; ic2 < childs.length; ic2++)
                 {
                     if(childs[ic2].className != 'creditsmissionlist') continue;
+                    if (gettypecredits(e.mtid) == undefined)
+                    {
+                        sessionStorage.clear("LSS_Missionrequirements")
+                        requirements = await getRequirements();
+                        sessionStorage.setItem("LSS_Missionrequirements", JSON.stringify(requirements));
+                        requirements = JSON.parse(sessionStorage.getItem("LSS_Missionrequirements"));
+                    }
                     let credits = gettypecredits(e.mtid).average_credits || 0;
 
                     gethtml_str(credits);
@@ -89,6 +96,13 @@
             else //create
             {
                 if(Missions[i].getAttribute('mission_type_id') == 'null') continue;
+                if (gettypecredits(Missions[i].getAttribute('mission_type_id')) == undefined)
+                {
+                    sessionStorage.clear("LSS_Missionrequirements")
+                    requirements = await getRequirements();
+                    sessionStorage.setItem("LSS_Missionrequirements", JSON.stringify(requirements));
+                    requirements = JSON.parse(sessionStorage.getItem("LSS_Missionrequirements"));
+                }
 
                 let credits = gettypecredits(Missions[i].getAttribute('mission_type_id')).average_credits || 0;
 
@@ -112,15 +126,27 @@
         // get  mission list
         var Missions = $('.missionSideBarEntry');
 
-        sessionStorage.clear("LSS_Missionrequirements")
-        requirements = await getRequirements();
-        sessionStorage.setItem("LSS_Missionrequirements", JSON.stringify(requirements));
-        requirements = JSON.parse(sessionStorage.getItem("LSS_Missionrequirements"));
+        if(sessionStorage.getItem("LSS_Missionrequirements") == null)
+        {
+            requirements = await getRequirements();
+            sessionStorage.setItem("LSS_Missionrequirements", JSON.stringify(requirements));
+        }
+        else
+        {
+            requirements = JSON.parse(sessionStorage.getItem("LSS_Missionrequirements"));
+        }
 
         // add info to mission
         for (var i = 0; i < Missions.length; i++)
         {
             if (Missions[i].getAttribute('mission_type_id') == 'null') continue;
+            if (gettypecredits(Missions[i].getAttribute('mission_type_id')) == undefined)
+            {
+                sessionStorage.clear("LSS_Missionrequirements")
+                requirements = await getRequirements();
+                sessionStorage.setItem("LSS_Missionrequirements", JSON.stringify(requirements));
+                requirements = JSON.parse(sessionStorage.getItem("LSS_Missionrequirements"));
+            }
 
             let credits = gettypecredits(Missions[i].getAttribute('mission_type_id')).average_credits || 0;
